+++ src/steps/IntoStep.ts
@@ -1,29 +1,51 @@
 import { BaseStep } from './BaseStep'
 import { BuilderData } from '../builder'
 import { PrimitiveType } from '../models/types'
-import { ValuesStep } from './ValuesStep'
-import { InsertColumnsAndValuesNotEqualError } from '../errors'
+import { DefaultValuesStep, ValuesStep } from './stepInterfaces'
+import { InsertColumnsAndExpressionsNotEqualError, InsertColumnsAndValuesNotEqualError } from '../errors'
 import { Binder } from '../binder'
+import { SelectStep } from './stepInterfaces'
+import { returnStepOrThrow } from '../util'
+import { SelectItemInfo } from '../SelectItemInfo'
+import { SelectItem } from './Step'
+import { Default } from '../singletoneConstants'
 
 export class IntoStep extends BaseStep {
   constructor(protected data: BuilderData) { super(data) }
 
-  public values(...values: (PrimitiveType|Binder)[]): ValuesStep {
-    this.throwIfColumnAndValueNotEqual(values)
+  public values(...values: (PrimitiveType|Binder|Default)[]): ValuesStep {
+    this.throwForInvalidValuesNumber(values)
     this.data.insertIntoValues.push(...values)
-    return new ValuesStep(this.data)
+    return returnStepOrThrow(this.data.step)
   }
 
   public values$(...values: PrimitiveType[]): ValuesStep {
-    this.throwIfColumnAndValueNotEqual(values)
+    this.throwForInvalidValuesNumber(values)
     this.data.insertIntoValues.push(...values.map(it => new Binder(it)))
-    return new ValuesStep(this.data)
+    return returnStepOrThrow(this.data.step)
   }
 
-  private throwIfColumnAndValueNotEqual(values: (PrimitiveType|Binder)[]) {
+  public defaultValues(): DefaultValuesStep {
+    this.data.insertIntoDefaultValues = true
+    return returnStepOrThrow(this.data.step)
+  }
+
+  public select(...items: (SelectItemInfo|SelectItem|PrimitiveType)[]): SelectStep {
+    this.throwForInvalidExpressionsNumber(items)
+    return returnStepOrThrow(this.data.step).select(...items)
+  }
+
+  private throwForInvalidValuesNumber(values: (PrimitiveType|Binder|Default)[]) {
     const columnsCount = this.data.insertIntoColumns.length
     if (columnsCount > 0 && columnsCount !== values.length) {
       throw new InsertColumnsAndValuesNotEqualError(columnsCount, values.length)
+    }
+  }
+
+  private throwForInvalidExpressionsNumber(items: (SelectItemInfo|SelectItem|PrimitiveType)[]) {
+    const columnsCount = this.data.insertIntoColumns.length
+    if (columnsCount > 0 && columnsCount !== items.length) {
+      throw new InsertColumnsAndExpressionsNotEqualError(columnsCount, items.length)
     }
   }
 }