+++ src/steps/Step.ts
@@ -20,8 +20,22 @@ import { BaseStep } from './BaseStep'
 import { SelectWhereStep } from './SelectWhereStep'
 import { HavingStep } from './HavingStep'
 import {
-  RootStep, SelectStep, SelectFromStep, CrossJoinStep, JoinStep, LeftJoinStep, RightJoinStep,
-  InnerJoinStep, FullOuterJoinStep, GroupByStep, OrderByStep, LimitStep, OffsetStep, ReturningStep,
+  CrossJoinStep,
+  DefaultValuesStep,
+  FullOuterJoinStep,
+  GroupByStep,
+  InnerJoinStep,
+  JoinStep,
+  LeftJoinStep,
+  LimitStep,
+  OffsetStep,
+  OrderByStep,
+  ReturningStep,
+  RightJoinStep,
+  RootStep,
+  SelectFromStep,
+  SelectStep,
+  ValuesStep,
 } from './stepInterfaces'
 import { LogicalOperator } from '../operators'
 import { FromItemRelation } from '../FromItemInfo'
@@ -36,15 +50,17 @@ export type ColumnLike = Column|Expression
 export type SelectItem = ColumnLike|AggregateFunction|Binder|Asterisk
 
 export class Step extends BaseStep
-  implements RootStep, SelectStep, SelectFromStep, CrossJoinStep, JoinStep, LeftJoinStep,
-    RightJoinStep, InnerJoinStep, FullOuterJoinStep, GroupByStep, OrderByStep, LimitStep, OffsetStep {
+  implements RootStep, SelectStep, SelectFromStep, CrossJoinStep, JoinStep, LeftJoinStep, RightJoinStep,
+    InnerJoinStep, FullOuterJoinStep, GroupByStep, OrderByStep, LimitStep, OffsetStep, ValuesStep, DefaultValuesStep {
   constructor(protected data: BuilderData) {
     super(data)
     data.step = this
   }
 
   public select(...items: (SelectItemInfo|SelectItem|PrimitiveType)[]): SelectStep {
-    this.data.sqlPath = SqlPath.SELECT
+    if (this.data.sqlPath === undefined) {
+      this.data.sqlPath = SqlPath.SELECT
+    }
     const selectItemInfos: SelectItemInfo[] = items.map(it => {
       if (it instanceof SelectItemInfo) {
         return it
@@ -65,13 +81,17 @@ export class Step extends BaseStep
   }
 
   public selectDistinct(...items: (SelectItemInfo|SelectItem|PrimitiveType)[]): SelectStep {
-    this.data.sqlPath = SqlPath.SELECT
+    if (this.data.sqlPath === undefined) {
+      this.data.sqlPath = SqlPath.SELECT
+    }
     this.data.distinct = ' DISTINCT'
     return this.select(...items)
   }
 
   public selectAll(...items: (SelectItemInfo|SelectItem|PrimitiveType)[]): SelectStep {
-    this.data.sqlPath = SqlPath.SELECT
+    if (this.data.sqlPath === undefined) {
+      this.data.sqlPath = SqlPath.SELECT
+    }
     this.data.distinct = ' ALL'
     return this.select(...items)
   }
@@ -244,6 +264,9 @@ export class Step extends BaseStep
   }
 
   public returning(...items: (ItemInfo|ReturningItem|PrimitiveType)[]): ReturningStep {
+    if (this.data.sqlPath === SqlPath.SELECT) {
+      throw new Error('Returning step can not be used in SELECT statement, It can be only use if the path start with INSERT, DELETE, or UPDATE')
+    }
     const returningItemInfo: ReturningItemInfo[] = items.map(it => {
       if (it instanceof ReturningItemInfo) {
         return it
@@ -271,11 +294,9 @@ export class Step extends BaseStep
     return this
   }
 
-  private throwIfColumnsNotInDb(columns: (ReturningItemInfo|SelectItemInfo|ColumnLike|Asterisk)[]) {
+  private throwIfColumnsNotInDb(columns: (ReturningItemInfo|SelectItemInfo|ColumnLike)[]) {
     for (const item of columns) {
-      if (item instanceof Asterisk) {
-        continue
-      } else if (
+      if (
         item instanceof Expression
         || item instanceof SelectItemInfo
         || item instanceof ReturningItemInfo