+++ test/exported/error.test.ts
@@ -21,8 +21,10 @@ import {
   NULLS_FIRST,
   NULLS_LAST,
   f,
+  ASTERISK,
 } from '../../src'
 import { database } from '../database'
+import { InsertColumnsAndExpressionsNotEqualError } from '../../src/errors'
 
 //Alias
 const ADD = ArithmeticOperator.ADD
@@ -338,14 +340,42 @@ describe('Throw desired Errors', () => {
 
       expect(actual).toThrow(InsertColumnsAndValuesNotEqualError)
     })
-    it(`values contain unsupported type`, () => {
-      const value = { unsupportedObject: 'something' }
+  })
+  describe('Error: InsertColumnsAndExpressionsNotEqualError', () => {
+    it(`columns more than expressions`, () => {
       function actual() {
-        // @ts-ignore
-        sql.insertInto(table1).values(value).getSQL()
+        sql.insertInto(table1, col1, col2).select('A')
       }
 
-      expect(actual).toThrow(`Value step has Unsupported value: ${value}, type: ${typeof value}`)
+      expect(actual).toThrow(InsertColumnsAndExpressionsNotEqualError)
     })
+    it(`expressions more than columns`, () => {
+      function actual() {
+        sql.insertInto(table1, col1, col2).select('A', 'B', 'C')
+      }
+
+      expect(actual).toThrow(InsertColumnsAndExpressionsNotEqualError)
+    })
+  })
+  it(`Throws error "Value step has Unsupported value: x, type: y"`, () => {
+    const value = { unsupportedObject: 'something' }
+    function actual() {
+      // @ts-ignore
+      sql.insertInto(table1).values(value).getSQL()
+    }
+
+    expect(actual).toThrow(`Value step has Unsupported value: ${value}, type: ${typeof value}`)
+  })
+  it(`Throws error "Insert statement must have values or select items"`, () => {
+    function actual() {
+      sql.insertInto(table1).getSQL()
+    }
+    expect(actual).toThrow(`Insert statement must have values or select items`)
+  })
+  it(`Throws error "Returning step can not be used in SELECT statement, It can be only use if the path start with INSERT, DELETE, or UPDATE"`, () => {
+    function actual() {
+      sql.selectAsteriskFrom(table1).returning(ASTERISK).getSQL()
+    }
+    expect(actual).toThrow(`Returning step can not be used in SELECT statement, It can be only use if the path start with INSERT, DELETE, or UPDATE`)
   })
 })